+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Features
 - Add `StudioClient` as connector to PhariaStudio for submitting traces.
+- You can now specify a `chunk_overlap` when creating an index in the Document Index.
 
 ### Fixes
 ...